refactor the additional parameters into the LLM wrapper
