+++ utils.js
@@ -143,7 +143,7 @@ export class LLMWrapper {
   #openAIKey;
   #googleKey;
   
-  model = new ModelCapabilities('gpt4o');
+  model = new ModelCapabilities('gpt-4o');
   openAIAPI = null;
 
   constructor(parameters) {
@@ -183,5 +183,54 @@ export class LLMWrapper {
             break;
     }
   }
+
+  static additionalParameters() {
+    const models = [ 
+        {label: "GPT-4o", value: 'gpt-4o'}, 
+        {label: "GPT-4o-mini", value: 'gpt-4o-mini'}, 
+        {label: "GPT-4.5-preview", value: 'gpt-4.5-preview'}, 
+        {label: "GPT-4.1", value: 'gpt-4.1'}, 
+        {label: "GPT-4.1-mini", value: 'gpt-4.1-mini'}, 
+        {label: "GPT-4.1-nano", value: 'gpt-4.1-nano'}, 
+        {label: "Gemini 2.5-flash", value: 'gemini-2.5-flash-preview-04-17'},
+        {label: "Gemini 2.5-pro", value: 'gemini-2.5-pro-preview-03-25'},
+        {label: "Gemini 2.0", value: 'gemini-2.0-flash'},
+        {label: "Gemini 2.0-Lite", value: 'gemini-2.0-flash-lite'},
+        {label: "Gemini 1.5", value: 'gemini-1.5-flash'},
+        {label: "o1", value: 'o1'}, 
+        {label: "o3-mini low", value: 'o3-mini low'}, 
+        {label: "o3-mini medium", value: 'o3-mini medium'}, 
+        {label: "o3-mini high", value: 'o3-mini high'},
+        {label: "o4-mini", value: 'o4-mini'}
+    ];
+
+    return [{
+            name: "openAIKey",
+            type: "string",
+            required: true,
+            uiElement: "password",
+            saveForUser: "global",
+            label: "Open AI API Key",
+            description: "Leave blank for the default, or your Open AI key - skprojectXXXXX"
+        },{
+            name: "googleKey",
+            type: "string",
+            required: true,
+            uiElement: "password",
+            saveForUser: "global",
+            label: "Google API Key",
+            description: "Leave blank for the default, or your Google API key - XXXXXX"
+        },{
+            name: "underlyingModel",
+            type: "string",
+            defaultValue: 'gpt-4o',
+            required: false,
+            options: models,
+            uiElement: "combobox",
+            saveForUser: "local",
+            label: "LLM Model",
+            description: "The LLM model that you want to use to process your queries."
+        }];
+    }
 };
